Handle corner cases properly while creating balanced KD-tree
Consider that the leaves on the lowest level are on the proper
side of the corresponding subtree depending on their values.
Especially one-sided leaves, ensure them to be on the left side
only in the case that their value of the corresponding
dimension less than or equal to their subtree root's value.

Avoid filling the same points into multiple nodes.